Remove redirect_from from yaml header
Whilst redirect_from seems to be creating the right files, with the right html in them, it seems that GitHub Pages serves up .php file in such a way that they are downloaded by the browser, rather than displayed.  So the user just winds up with a short .php file in their downloads folder, not on the page that we want them to see.  Probably better to do a custom "404: File Not Found" error page.